HevSocks5SessionTCP: Fix use after free.
This issue occurs after `tcp_close` is called, where operations driven by
the TCP timer may still access the TCP data buffer. In the zero-copy case,
this would make it possible to access the freed ring buffer.

The way to continue using zero-copy and resolve this issue is to invoke
`tcp_abort` to forcefully terminate the TCP connection. Before terminating
the connection, it should wait for a maximum timeout period to clear the
send buffer in order to prevent data loss.

Fixes: https://github.com/heiher/hev-socks5-tunnel/issues/74
Reported-by: Yewenyu <289193866@qq.com>
Signed-off-by: hev <r@hev.cc>